==> cohort_analysis/2_DataCleaning.ipynb <==
import pandas as pd
import numpy as np
from pandas.api.types import CategoricalDtype

Dictionary Creation

race_dict = {
    ' Other': 'Unknown',
    ' Unknown': 'Unknown',
    ' Patient Refused': 'Unknown',
    ' Other ~ Unknown': 'Unknown',

    ' White or Caucasian': 'White',
    ' Other ~ White or Caucasian': 'White',
    ' Patient Refused ~ White or Caucasian': 'White',
    ' Unknown ~ White or Caucasian': 'White',

    ' Black or African American': 'Black or African American',
    ' Black or African American ~ Other': 'Black or African American',
    ' Black or African American ~ Unknown': 'Black or African American',
    ' Black or African American ~ Other ~ Unknown': 'Black or African American',

    ' Asian ~ Unknown': 'Asian',
    ' Asian': 'Asian',
    ' Asian ~ Asian Indian': 'Asian',
    ' Asian ~ Filipino': 'Asian',
    ' Nepalese ~ Other': 'Asian',
    ' Pakistani': 'Asian',
    ' Nepalese': 'Asian',
    ' Asian Indian ~ Unknown': 'Asian',
    ' Bhutanese ~ Other': 'Asian',
    ' Asian ~ Chinese': 'Asian',
    ' Asian ~ Other': 'Asian',
    ' Indonesian': 'Asian',
    ' Asian Indian': 'Asian',
    ' Laotian ~ Other': 'Asian',
    ' Asian ~ Vietnamese': 'Asian',

    ' Other Pacific Islander': 'Native Hawaiian or Other Pacific Islander',
    ' Guamanian': 'Native Hawaiian or Other Pacific Islander',

    ' American Indian or Alaskan Native': 'American Indian or Alaska Native',   
    ' American Indian or Alaskan Native ~ Other': 'American Indian or Alaska Native',

#"More than One Race" See the highest % race assignment for Table 4 in https://wonder.cdc.gov/wonder/help/populations/bridged-race/PublicHealthReports119-2-p192.pdf
# see also https://doh.wa.gov/sites/default/files/legacy/Documents/1500/RaceEthnGuidelines.pdf
# see also https://www.cdc.gov/nchs/data/dvs/Multiple_race_documentation_5-10-04.pdf

    ' Palauan ~ White or Caucasian': 'More than one race',
    ' Bangladeshi ~ White or Caucasian': 'More than one race',
    ' American Indian or Alaskan Native ~ White or Caucasian': 'More than one race',
    ' Asian Indian ~ White or Caucasian': 'More than one race',
    ' White or Caucasian ~ Yapese': 'More than one race',
    ' Asian ~ White or Caucasian': 'More than one race',
    ' Black or African American ~ White or Caucasian': 'More than one race',
    ' American Indian or Alaskan Native ~ Black or African American': 'More than one race',
    ' Black or African American ~ Native Hawaiian': 'More than one race',
    ' Black or African American ~ Indonesian ~ Other': 'More than one race',
    ' Black or African American ~ Other ~ White or Caucasian': 'More than one race',
}

race_dict_lkp = {
    'Unknown': 0,
    'American Indian or Alaska Native': 1,
    'Asian': 2,
    'Black or African American': 3,
    'Native Hawaiian or Other Pacific Islander': 4,
    'White': 5,
    'More than one race': 6,
}


ethnicity_dict = {
    ' Unknown': 'Unknown',
    ' Patient Refused': 'Unknown',
    ' Patient Refused ~ Unknown': 'Unknown',
    ' Hispanic or Latino ~ Not Hispanic or Latino': 'Unknown',

    ' Not Hispanic or Latino': 'Not Hispanic',
    ' Not Hispanic or Latino ~ Unknown': 'Not Hispanic',
    ' Not Hispanic or Latino ~ Patient Refused': 'Not Hispanic',
    ' Not Hispanic or Latino ~ Uruguayan': 'Not Hispanic',

    ' Hispanic or Latino': 'Hispanic',
    ' Puerto Rican': 'Hispanic',
    ' Mexican American Indian': 'Hispanic',
    ' Hispanic or Latino ~ Unknown': 'Hispanic',
    ' Spaniard ~ Unknown': 'Hispanic',
    ' Hispanic or Latino ~ Puerto Rican': 'Hispanic',
    ' Peruvian': 'Hispanic',
    ' Guatemalan ~ Honduran ~ Puerto Rican ~ Spaniard': 'Hispanic',
    ' Central American': 'Hispanic',
    ' Spaniard': 'Hispanic',
    ' Puerto Rican ~ Unknown': 'Hispanic',
    ' Dominican ~ Puerto Rican': 'Hispanic',
    ' Mexican American': 'Hispanic',
    ' Guatemalan': 'Hispanic',
    ' South American Indian': 'Hispanic',
    ' Central American Indian': 'Hispanic'
}

ethnicity_dict_lkp = {
    'Hispanic': 1,
    'Not Hispanic': 2,
}

insurance_dict = {
    'Commercial': 'Private',
    'Out of Area BC/BS': 'Private',
    'Government Other': 'Government',
    'Medicare': 'Medicare',
    'Medicaid': 'Medicaid',
    'Medicare Advantage': 'Medicare',
    'Medicaid Managed Care': 'Medicaid',
    'Excellus': 'Private',
    'MVA': 'Government',
    'Aetna': 'Private',
    'MVP': 'Private',
    "Worker's Comp": 'Government',
    'Institutional': 'Private',
    'UNIVERA SENIOR CHOICE MEDICARE': 'Medicare',
    'MEDICARE PART A AND B': 'Medicare',
    'MVP PREMIER INDIVIDUAL': 'Private',
    'EXCELLUS': 'Private',
    'UNITED HEALTHCARE MEDICAID': 'Medicaid'
}

insurance_dict_lkp = {
    'Private': 1,
    'Medicare': 2,
    'Medicaid': 3,
    'Government': 4
}

assessment_color_dict = {
    'RED': 'Red',
    'YELLOW': 'Yellow',
    'BLUE': 'Blue'
}

assessment_color_dict_lkp = {
    'Red': 1,
    'Yellow': 2,
    'Blue': 3
}

discharge_status_dict_lkp = {
    'Home or Self Care': 1,
    'Patient Expired': 2,
    'To Home Health Org Care': 3,
    'To Jail / Law Enforcement Facility': 4,
    'To SNF (Skilled Nursing)': 5,
    'To Inpatient Rehab Facility or Unit': 6,
    'Left against medical advice': 7,
    'To Hospice/Home Care': 8,
    'To Psychiatric Hospital or Unit': 9,
    'To Short Term Acute Care Hosp': 10,
    'To Hospice/Medical Facility': 11,
    'To LTC Facility (Long Term Care)': 12,
    'Sent to SMH': 13,
    'Sent to HH': 14,
    'To Short Term General Hospital for Inpatient Care with Planned Hospital Readmission': 15,
    'To Inpatient Rehab Facility or Unit with Planned Hospital Readmission': 16,
    'To Other Facility not otherwise defined': 17,
    'To ICF (Intermediate Care)': 18,
    'To Federal Hospital': 19,
    "To Designated Cancer Ctr or Children's Hospital with Planned Hospital Readmission": 20,
    'Still Inpatient': 21,
    'To Psychiatric Hospital or Unit with Planned Hospital Readmission': 22
}

covid_dict = {
    'Yes': 1,
    'No': 0
}

covid_dict_lkp = {
    'Positive': 1,
    'Negative': 0
}

sex_dict = {
    'M': 'Male',
    'F': 'Female'
}


sex_dict_lkp = {
    'Male': 1,
    'Female': 2
}

#Reduced Race Dictionary
def race_ethnicity_dict (row):
    if row['Ethnicity'] == 'Hispanic' :
        return 'Hispanic'
    return row['Race']

Extract and re-format data

#Extract data from csv, package in df's*, and convert encounter numbers to 
study_cohort_df = pd.read_excel('VAP_DeID_2022-01-12.xlsx', sheet_name='Study_Cohort_DeID', converters={'Encounter_Number': '{:0>4}'.format, 'SubjectID':str})
insurance_xl_df = pd.read_excel('VAP_DeID_2022-01-12.xlsx', sheet_name='Insurance_DeID', converters={'Encounter_Number': '{:0>4}'.format, 'SubjectID':str})
covid_df = pd.read_excel('VAP_DeID_2022-01-12.xlsx', sheet_name='COVID_Status_DeID', converters={'SubjectID':str})
blis_df = pd.read_excel('VAP_DeID_2022-01-12.xlsx', sheet_name='BLIS_ASSESSMENT_DeID', converters={'Encounter_Number': '{:0>4}'.format, 'SubjectID':str})
admission_df = pd.read_excel('VAP_DeID_2022-01-12.xlsx', sheet_name='Admission_dXs_DeID', converters={'Encounter_Number': '{:0>4}'.format, 'SubjectID':str})

#Create lookup dataframes#
sex_lkp_df = pd.DataFrame(sex_dict_lkp.items(), columns=['Description', 'ID'])
race_lkp_df = pd.DataFrame(race_dict_lkp.items(), columns=['Description', 'ID'])
insurance_lkp_df = pd.DataFrame(insurance_dict_lkp.items(), columns=['Description', 'ID'])
ethnicity_lkp_df = pd.DataFrame(ethnicity_dict_lkp.items(), columns=['Description', 'ID'])
covid_lkp_df = pd.DataFrame(covid_dict_lkp.items(), columns=['Description', 'ID'])
assessment_color_lkp_df = pd.DataFrame(assessment_color_dict_lkp.items(), columns=['Description', 'ID'])
discharge_status_lkp_df = pd.DataFrame(discharge_status_dict_lkp.items(), columns=['Description', 'ID'])

#Create ICD_10 lookup
#Create ICD_10 lookup
icd10_lkp_df = (
    admission_df[['ICD10_dX(s)', 'dX_Name']]
    .copy()
    .drop_duplicates()
    .reset_index(drop=True)
    .dropna(subset = {'ICD10_dX(s)'})
)
icd10_lkp_df['ID'] = icd10_lkp_df.index
icd10_lkp_df = icd10_lkp_df[['ID', 'ICD10_dX(s)', 'dX_Name']]
icd10_lkp_df.columns = ['ID', 'Code', 'Description']

#swap columns in LkP dataframes#
sex_lkp_df = sex_lkp_df[['ID', 'Description']]
race_lkp_df = race_lkp_df[['ID', 'Description']]
insurance_lkp_df = insurance_lkp_df[['ID', 'Description']]
ethnicity_lkp_df = ethnicity_lkp_df[['ID', 'Description']]
covid_lkp_df = covid_lkp_df[['ID', 'Description']]
assessment_color_lkp_df = assessment_color_lkp_df[['ID', 'Description']]
discharge_status_lkp_df = discharge_status_lkp_df[['ID', 'Description']]

#apply dictionaries
study_cohort_df['Race'] = study_cohort_df['Race'].map(race_dict).map(race_dict_lkp)
study_cohort_df['Ethnicity'] = study_cohort_df['Ethnicity'].map(ethnicity_dict).map(ethnicity_dict_lkp)#.astype(CategoricalDtype(categories={1:'Hispanic', 2:'Not Hispanic'})).cat.codes
study_cohort_df['Sex'] = study_cohort_df['Sex'].map(sex_dict).map(sex_dict_lkp)#.astype(CategoricalDtype(categories={1:'Male', 2:'Female'})).cat.codes
study_cohort_df['Discharge_Status'] = study_cohort_df['Discharge_Status'].map(discharge_status_dict_lkp)
study_cohort_df['Reduced_Race'] = study_cohort_df.apply(race_ethnicity_dict, axis=1)
insurance_xl_df['Insurance'] = insurance_xl_df['Insurance'].map(insurance_dict).map(insurance_dict_lkp)
covid_df['COVID_POSITIVE_N3C_Phenotype'] = covid_df['COVID_POSITIVE_N3C_Phenotype'].map(covid_dict)
blis_df['assessment_color'] = blis_df['assessment_color'].map(assessment_color_dict).map(assessment_color_dict_lkp)

print(blis_df.dtypes)
print(study_cohort_df.info())

study_cohort_df.to_pickle("cohort_raw.pkl")
admission_df.to_pickle("diagnoses_raw.pkl")

## Data Cleaning

#included #1: Create list of SubjectIDs that have rows and relevant info in ALL of subjects_df, covid_df, insurance_xl_df, blis_df and admission_df
included_1 = (study_cohort_df
    .merge(covid_df[["SubjectID", "COVID_POSITIVE_N3C_Phenotype"]], on="SubjectID") #identify all subjects w/ COVID records
    .merge(insurance_xl_df[["SubjectID", "Insurance"]], on="SubjectID")
    .merge(blis_df[["SubjectID", "intubation_number"]], on="SubjectID")
    .merge(admission_df[["SubjectID", "Encounter_Number", "ICD10_dX(s)"]], on=["SubjectID","Encounter_Number"])
    .dropna(subset=['Discharge_Status', "ICD10_dX(s)", "Age_at_Admission"]) #drop records that do not have discharge, comorbidity, or age at admission data.
    .drop_duplicates(subset=['SubjectID', 'Encounter_Number'])
    .reindex(columns=['SubjectID', 'Encounter_Number'])
)

#included #2: Identify SubjectIDs that have a 0 assessment timepoint for intubation #1
included_2 = (blis_df
    .where(blis_df["intubation_number"]==1)
    .where(blis_df["assessment_timepoint"]==0)
    .dropna()
    .drop_duplicates('SubjectID')
    .reindex(columns=['SubjectID'])
)

#Error #3: Identify SubjectIDs where: 
#(1) There is a BLIS record AND 
#(2a) # of Encounters = 1 OR (2b) # of Encounters = # of Intubations 

encounter_counts= (study_cohort_df
    .merge(study_cohort_df['SubjectID']
            .value_counts()
            .rename_axis('SubjectID')
            .reset_index(name='Encounter_Count')
    )
    .reindex(columns=[
        'SubjectID',
        'Encounter_Count']
        )
)

intubation_counts= (blis_df
    .groupby('SubjectID')
    .max()
    .rename(columns = {'intubation_number': 'Intubation_Count'})
    .reset_index('SubjectID')
    .reindex(columns=[
        'SubjectID',
        'Intubation_Count']
        )
)

included_3_merge = (study_cohort_df
    .merge(encounter_counts, on="SubjectID")
    .merge(intubation_counts, on="SubjectID")
    .drop_duplicates()
)
included_3 = (included_3_merge    
    .where((included_3_merge['Encounter_Count']==1) | (included_3_merge['Encounter_Count'] == included_3_merge['Intubation_Count']))
    .dropna(subset=['SubjectID'])
    .drop_duplicates('SubjectID')
    .reindex(columns=['SubjectID'])#, 'Encounter_Number'])
)

#Compile full list of error-free SubjectIDs
included = (included_1
    .merge(included_2)
    .merge(included_3)
    #.rename(columns={'Encounter_Number':"EncounterID"})
)

###CREATE Encounter-to-Intubation crosswalks ###

#create Encounter-to-Intubation crosswalk for all SubjectIDs with 1 Encounters
vent_encounter_1_df = (
    included_3_merge
    .where(included_3_merge['Encounter_Count']==1) #select all Subject_IDs where only 1 encounter
    .rename(columns={'Encounter_Number':'EncounterID'})
    .assign(encounter_number = 1) #create column 'encounter_number' and assign value of 1 to all rows
    .loc[included_3_merge.index.repeat(included_3_merge['Intubation_Count'])]     #copy each row N times, where N is value in 'Intubation_Count'
    )
vent_encounter_1_df = (
    vent_encounter_1_df
    .assign(intubation_number = vent_encounter_1_df.groupby(['SubjectID', 'EncounterID']).cumcount()+1) #assign ordering of intubations by 'SubjectID' and 'EncounterID'
    .dropna(subset=['SubjectID'])
    .drop_duplicates()
    .reset_index()
)

#create Encounter-to-Intubation crosswalk for all SubjectIDs with <1 Encounters
vent_encounter_multi_df = (
    included_3_merge
    .where((included_3_merge['Encounter_Count'] !=1) &
            (included_3_merge['Encounter_Count'] == included_3_merge['Intubation_Count']))
    .rename(columns={'Encounter_Number':'EncounterID'})
)
vent_encounter_multi_df = (
    vent_encounter_multi_df
    .assign(encounter_number = vent_encounter_multi_df.groupby(['SubjectID']).cumcount()+1)
    .assign(intubation_number = included_3_merge.groupby(['SubjectID']).cumcount()+1) #assign ordering of intubations within subject-encounter pairs
    .dropna(subset=['SubjectID'])
    .drop_duplicates()
)

#Concat _1 and _multi frames to create full Encounter-to-Intubation crosswalk
vent_encounter_df = pd.concat([vent_encounter_1_df, vent_encounter_multi_df]) #add rows from vent_encounter_multi to bottom of vent_encounter_1
vent_encounter_df = (
    vent_encounter_df
    .reindex(columns = ['SubjectID', 
            'EncounterID',
            'intubation_number',
            'Encounter_Count',
            'Discharge_Status',
            ]
        )
    .reset_index(drop=True)
    )

#generate list of encounters where ventilation was likely due to elective procedure.
elective_flag = (admission_df
      .where(admission_df["ICD10_dX(s)"].str.contains("Z01.8"))
      .dropna(subset=['ICD10_dX(s)'])
      .assign(Elective_Flag=1)
      .rename(columns={'Encounter_Number':'EncounterID'})
      .reindex(columns={'EncounterID', 'Elective_Flag'})
      .drop_duplicates()
      )

### Calculate Life Expectancy and Comorbidity Adjusted Triage Scores

## Calculate:
# (1) Life Expectancy and Comorbidity-Adjusted Life Expectancy
# (2) Colorado Crisis Standards of Care Score
# (3) Maryland Multi-Principle Score

%run 2a_LifeExpCalc.ipynb

## Dataframe creation

#create subject dataframe
subject_df = (study_cohort_df[['SubjectID', 'Race', 'Ethnicity', 'Sex']]
    .merge(included) #remove all error SubjectIDs
    .drop(columns="Encounter_Number")
    .drop_duplicates()
    .sort_values(by="SubjectID")
    .reset_index(drop=True)
)

#create encounter dataframe
encounter_df = (study_cohort_df[['SubjectID', 'Encounter_Number', 'Discharge_Status', 'Length_of_Stay (days)', 'Age_at_Admission']]
    .merge(included) #remove all error SubjectIDs
    .merge(covid_df[["SubjectID", "COVID_POSITIVE_N3C_Phenotype"]], on="SubjectID")
    .merge(encounter_le[['Encounter_Number', 'Life_Exp', 'Cho_Tier', 'Cho_LE']], on = "Encounter_Number")   
    .rename(columns={
        'Encounter_Number': 'EncounterID', 
        'Length_of_Stay (days)': 'Length_of_Stay_Days', 
        'COVID_POSITIVE_N3C_Phenotype': 'COVID_Status'
        })
    .merge(encounter_comorbid, on='EncounterID') #bring in Charlson and Elixhauser scores
    .merge(elective_flag, on="EncounterID", how="left") #bring in flag for elective procedures
    .drop_duplicates()
    .fillna({'Elective_Flag':0}) #fill missing electives flags with 0 (i.e. non-elective)  
    .sort_values(by="EncounterID")
    .reindex(columns=[
        'EncounterID',
        'SubjectID',
        'Discharge_Status',
        'Length_of_Stay_Days',
        'Age_at_Admission',
        'COVID_Status',
        'Elective_Flag',
        'CCS_raw',
        'CCS_age',
        'CCS_Colorado',
        'ECI_raw',
        'Life_Exp', 
        'Cho_Tier', 
        'Cho_LE']
        )
    .reset_index(drop=True)
)

# create insurance dataframe
insurance_df = (insurance_xl_df[['SubjectID', 'Encounter_Number', 'Insurance']]
    .merge(included) #remove all error SubjectIDs
    .drop_duplicates(subset=['SubjectID', 'Encounter_Number', 'Insurance'])
    .rename(columns={'Encounter_Number': 'EncounterID', 'Insurance':'Insurance_CD'})
    .sort_values(by=["SubjectID", "EncounterID"])
    .reindex(columns=[
        'ID',
        'SubjectID',
        'EncounterID',
        'Insurance_CD',]
        )
#    .set_index(['SubjectID', 'EncounterID'])
    .reset_index(drop=True)
)

#create comorbidity dataframe
comorbid_df = (admission_df[['SubjectID', 'Encounter_Number', 'ICD10_dX(s)', 'Reason_Visit']]
    .merge(included)#, on="SubjectID")
    .drop_duplicates(subset=['SubjectID', 'Encounter_Number', 'ICD10_dX(s)'])
    .rename(columns={'Encounter_Number': 'EncounterID', 'ICD10_dX(s)':'ICD_10'})
    .sort_values(by=["SubjectID", "EncounterID"])
    .reindex(columns=[
        'ID',
        'SubjectID', 
        'EncounterID',
        'ICD_10',
        'Reason_Visit']
        )
    .reset_index(drop=True)
)

#create ventilator table
vent_df = (blis_df
    .merge(included, on=['SubjectID']) #remove all error SubjectIDs
    .merge(vent_encounter_df, on=['SubjectID', 'intubation_number'])  #associate Encounter #s with SubjectIDs & Intubations
    .merge(encounter_comorbid, on='EncounterID')
    .assign()
    .assign(PA_S = lambda df_: df_.apply(PA_sofa_calc, axis=1),
            Colorado_S = lambda df_: df_.apply(colorado_sofa_calc, axis=1),
            Maryland_S = lambda df_: df_.apply(Maryland_sofa_calc, axis=1)
        )
    .assign(seq_num = lambda vent_df_y: #create a unique ID for each row
                vent_df_y['EncounterID'].map(str) + '-' + 
                vent_df_y['intubation_number'].map(str) + '-' + 
                vent_df_y['assessment_timepoint'].astype(str).str.zfill(3),
            intubation_number = lambda vent_df_y: #assign intubation_number = 1 for all encounters where Encounter_Count >1
                vent_df_y['intubation_number'].mask(vent_df_y.Encounter_Count >= 2, 1),
            Maryland_Score = lambda df_y: df_y['Maryland_S']+df_y['Maryland_C'],
            PA_Score = lambda df_y: df_y['PA_S']+df_y['PA_C'],
            Colorado_Score = lambda df_y: df_y['Colorado_S']+df_y['Colorado_C']
        )
    .drop_duplicates(subset=["seq_num"])
    .rename(columns={
        'seq_num':'EIT',
        'intubation_number':'Intubation',
        'assessment_timepoint':'Timepoint',
        'vent_duration (hours)':'Vent_Duration',
        'sofa_score':'SOFA',
        'assessment_color':'NY_Score'}
        )
    .reindex(columns=[
        'EIT',
        'EncounterID',
        'SubjectID',
        'Intubation',
        'Timepoint',
        'Vent_Duration',
        'SOFA',
        'NY_Score',
        'Maryland_Score',
        'Colorado_Score',
        'PA_Score'
        ])
    .sort_values(by=["EIT"])
)


## Data Analysis of Excluded vs Included Subjects


##Check for size

check_df = (vent_df 
    .merge(encounter_df, on = ['EncounterID', 'SubjectID'])
    .merge(comorbid_df, on = ['EncounterID', 'SubjectID'])
    .dropna(subset=["ICD_10"])
    .reindex(columns=['EncounterID', 'SubjectID'])
    .drop_duplicates()
)

print('Number of Unique Encounters: ', check_df['EncounterID'].drop_duplicates().count())
print('Number of Unique Subjects: ', check_df['SubjectID'].drop_duplicates().count())
print(check_df.info())

print('Number of Unique Encounters: ', study_cohort_df['Encounter_Number'].drop_duplicates().count())
print('Number of Unique Subjects: ', study_cohort_df['SubjectID'].drop_duplicates().count())
print(study_cohort_df.info())


database_pull_df = pd.read_pickle("encounters.pkl")
print('Number of Unique Encounters: ', database_pull_df['EncounterID'].drop_duplicates().count())
print('Number of Unique Subjects: ', database_pull_df['SubjectID'].drop_duplicates().count())
print(database_pull_df.info())

check_2 = check_df.merge(database_pull_df, on=["SubjectID", "EncounterID"], how="outer", indicator=True)
check_2 = check_2[check_2._merge != 'both']

#Compile list of error SubjectIDs (i.e. excluded)
excluded_IDs = (study_cohort_df
    .rename(columns = {"Encounter_Number": "EncounterID"})
    .reindex(columns = ["SubjectID", "EncounterID"])
    .merge(database_pull_df, how='outer', indicator=True)
    .rename(columns = {"EncounterID":"Encounter_Number"})
#    .reindex(columns=['EncounterID', '_merge'])
)
excluded_IDs = excluded_IDs[excluded_IDs._merge != 'both'].reindex(columns=['SubjectID', 'Encounter_Number'])


excluded_df = (study_cohort_df#[['SubjectID', 'Encounter_Number', 'Discharge_Status', 'Length_of_Stay (days)', 'Age_at_Admission']]
    .merge(excluded_IDs, on=["SubjectID", "Encounter_Number"])
    .merge(covid_df[["SubjectID", "COVID_POSITIVE_N3C_Phenotype"]], on="SubjectID", how='left')
    .merge(encounter_le[['Encounter_Number','Life_Exp','Cho_Tier', 'Cho_LE']], on='Encounter_Number', how="left") #bring in Life Expectancy and Cho-Adjusted LE
    .rename(columns={ 
        "Encounter_Number": 'EncounterID',
        'Length_of_Stay (days)': 'StayLength', 
        'COVID_POSITIVE_N3C_Phenotype': 'COVID_Status',
        'Age_at_Admission': 'Age'
        })
    .sort_values(by="EncounterID")
    .merge(encounter_comorbid[['EncounterID', 'CCS_raw', 'CCS_age', 'CCS_Colorado','ECI_raw']], on='EncounterID', how="left") #bring in Charlson and Elixhauser scores
    .reindex(columns=[
        'EncounterID',
        'SubjectID',
        'Age',
        'Race',
        'Ethnicity',
        'Sex',
        'InitialSOFA', #Column is entirely NaN, but needs to exist for processing by calculators.ipynb
        'StayLength',
        'Discharge_Status',
        'COVID_Status',
        'CCS_raw',
        'CCS_age',
        'CCS_Colorado',
        'ECI_raw',
        'Life_Exp', 
        'Cho_Tier', 
        'Cho_LE']
        )
    .drop_duplicates()

)

print (excluded_df.info())

excluded_df.to_pickle("excluded_raw.pkl")
